pyproject.toml
[tool.poetry]
name = "aiohttp-graphql"
version = "0.1.0"
description = ""
authors = ["Tiago Requeijo <tiago.requeijo.dev@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
graphene = {version = ">=3.0.dev20190817210753",optional = true,allows-prereleases = true}
graphql-core = {version = ">=3.0.0b1",allows-prereleases = true}
aiohttp = "^3.6"

[tool.poetry.dev-dependencies]
pytest = "^5.0"
pytest-cov = "^2.6"
pytest-flake8 = "^1.0"
pytest-mypy = "^0.4.1"
flake8-docstrings = "^1.3"
flake8-comprehensions = "^2.0"
flake8-import-order = "^0.18.0"
flake8-bugbear = "^18.8"
flake8-blind-except = "^0.1.1"
flake8-builtins = "^1.4"
pytest-black = "^0.3.7"
pytest-asyncio = "^0.10.0"

[tool.poetry.extras]
graphene = ["graphene"]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
